Trimmed rtu v2 examples
