+++ ScanbotBarcodeSDKExample/BarcodeSDK.NET.Droid.Example/Activities/MainActivity.RTUv2.cs
@@ -10,6 +10,153 @@ namespace BarcodeSDK.NET.Droid
 {
     public partial class MainActivity : Activity
     {
+        private void SingleScanning(object sender, EventArgs e)
+        {
+            if (!Alert.CheckLicense(this, SDK))
+            {
+                return;
+            }
+            
+            var intent = BarcodeScannerActivity.NewIntent(this, new BarcodeScannerConfiguration
+            {
+                RecognizerConfiguration = new BarcodeRecognizerConfiguration
+                {
+                    BarcodeFormats = BarcodeFormat.Values().ToList(),
+                    Gs1Handling = Gs1Handling.Decode
+                },
+                UseCase = new SingleScanningMode()
+            });
+            StartActivityForResult(intent, BARCODE_DEFAULT_UI_REQUEST_CODE_V2);
+        }
+
+        private void SingleScanningWithArOverlay(object sender, EventArgs e)
+        {
+            if (!Alert.CheckLicense(this, SDK))
+            {
+                return;
+            }
+            
+            var intent = BarcodeScannerActivity.NewIntent(this, new BarcodeScannerConfiguration
+            {
+                UseCase = new SingleScanningMode()
+                {
+                    ArOverlay = new ArOverlayGeneralConfiguration() 
+                    {
+                        Visible = true
+                    }    
+                }
+            });
+            StartActivityForResult(intent, BARCODE_DEFAULT_UI_REQUEST_CODE_V2);
+        }
+
+        private void BatchBarcodeScanning(object sender, EventArgs e)
+        {
+            if (!Alert.CheckLicense(this, SDK))
+            {
+                return;
+            }
+            
+            var intent = BarcodeScannerActivity.NewIntent(this, new BarcodeScannerConfiguration 
+            {
+                RecognizerConfiguration = new BarcodeRecognizerConfiguration
+                {
+                    BarcodeFormats = BarcodeFormat.Values().ToList(),           
+                },
+                UseCase = new MultipleScanningMode
+                {
+                    Mode = MultipleBarcodesScanningMode.Counting
+                }
+            });
+            StartActivityForResult(intent, BARCODE_DEFAULT_UI_REQUEST_CODE_V2);
+        }
+
+        private void MultipleUniqueBarcodeScanning(object sender, EventArgs e)
+        {
+            if (!Alert.CheckLicense(this, SDK))
+            {
+                return;
+            }
+            
+            var intent = BarcodeScannerActivity.NewIntent(this, new BarcodeScannerConfiguration
+            {
+                UseCase = new MultipleScanningMode
+                {
+                    Mode = MultipleBarcodesScanningMode.Counting
+                },
+                UserGuidance = new UserGuidanceConfiguration
+                {
+                    Title = new StyledText{ Text = "Please align the QR-/Barcode in the frame above to scan it." }
+                }
+            });
+            StartActivityForResult(intent, BARCODE_DEFAULT_UI_REQUEST_CODE_V2);
+        }
+
+        private class CustomMapper : global::Java.Lang.Object, global::Java.IO.ISerializable, IO.Scanbot.Sdk.Ui_v2.Barcode.Configuration.IBarcodeItemMapper
+        {
+            public void MapBarcodeItem(BarcodeItem barcodeItem, IBarcodeMappingResult result)
+            {
+                var title = $"Some product {barcodeItem.TextWithExtension}";
+                var subTitle = barcodeItem.Type.Name();
+                var image = "https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png";
+
+                if (barcodeItem.TextWithExtension == "Error occurred!")
+                {
+                    result.OnError();
+                }
+                else
+                {
+                    result.OnResult(new BarcodeMappedData(title: title, subtitle: subTitle, barcodeImage: image));
+                }
+            }
+        }
+
+        private void FindAndPickScanning(object sender, EventArgs e)
+        {
+            if (!Alert.CheckLicense(this, SDK))
+            {
+                return;
+            }
+
+            var configuration = new BarcodeScannerConfiguration();
+
+            // Initialize the use case for multiple scanning.
+            var findAndPickConfig = new FindAndPickScanningMode();
+
+            // Set the sheet mode for the barcodes preview.
+            findAndPickConfig.Sheet.Mode = SheetMode.CollapsedSheet;
+
+            // Enable/Disable the automatic selection.
+            findAndPickConfig.ArOverlay.AutomaticSelectionEnabled = false;
+
+            // Set the height for the collapsed sheet.
+            findAndPickConfig.Sheet.CollapsedVisibleHeight = CollapsedVisibleHeight.Large;
+
+            // Enable manual count change.
+            findAndPickConfig.SheetContent.ManualCountChangeEnabled = true;
+
+            // Set the delay before same barcode counting repeat.
+            findAndPickConfig.CountingRepeatDelay = 1000;
+
+            // Configure the submit button.
+            findAndPickConfig.SheetContent.SubmitButton.Text = "Submit";
+            
+            findAndPickConfig.SheetContent.SubmitButton.Foreground.Color = new ScanbotColor("#000000"); //arg string
+
+            // Set the expected barcodes.
+            // findAndPickConfig.ExpectedBarcodes = new List<ExpectedBarcode>() 
+            // {
+            //     new ExpectedBarcode(barcodeValue: "123456", title: "numeric barcode", image: "https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png", count: 4),
+            //     new ExpectedBarcode(barcodeValue: "SCANBOT", title: "value barcode", image: "https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png", count: 4),
+            // };
+
+            // Configure other parameters, pertaining to findAndPick-scanning mode as needed.
+            configuration.UseCase = findAndPickConfig;
+            configuration.RecognizerConfiguration.BarcodeFormats = BarcodeFormat.Values().ToList();
+
+            var intent = BarcodeScannerActivity.NewIntent(this, configuration);
+            StartActivityForResult(intent, BARCODE_DEFAULT_UI_REQUEST_CODE_V2);
+        }
+
         private void OnRTUv2ActivityResult(Intent data, BarcodeScannerResult barcode)
         {
             var imagePath = data.GetStringExtra(
